debug: client lint 수정
