+++ client/src/vite-env.d.ts
@@ -1,1 +1,2 @@
+// eslint-disable-next-line spaced-comment
 /// <reference types="vite/client" />
